Source for abcl code has it's location recorded as in the jar file. However in most cases we want to edit in the source tree. This patch modifies find-definitons to add the source tree information if it is added to *source-path*. To do that, setf (get :swank-abcl-source-path :path) to a list of the absolute path to "src" in the abcl local repo (for ABCL proper) and the absolute path to the abcl repo (for contrib).
Variable *when-in-file-system-and-jar-choose* lets you control behavior
in this case.  Set to :both to have the choice, :filesystem for just the
filesystem or :jar for just the jar.

In order to do this some surgery needed to be done. Also,
some code was rearranged so that the source location
code was closer together.